requirements.txt
flask
google-cloud-bigquery
openai
python-dotenv
functions-framework==3.*